fix: code refactor
