+++ styles/globals.scss
@@ -66,9 +66,9 @@
   padding: 12px !important;
 }
 
-// .slick-slide {
-//   width: 320px !important;
-// }
+.slick-slide {
+  width: 320px !important;
+}
 
 // Navbar
 .top-navbar {
@@ -189,8 +189,4 @@
 .navbar_category {
   text-align: right !important;
   color: #cd8825 !important;
-}
-// top arrow carousel
-.top-arrow-carousel .slick-slide {
-  width: 100% !important;
 }
